// convex/_generated/api.ts
export const api = {
// ...
  projects: {
    listByOwner: {} as FunctionReference<"query">,
    get: {} as FunctionReference<"query">,
    create: {} as FunctionReference<"mutation">,
    update: {} as FunctionReference<"mutation">,
    remove: {} as FunctionReference<"mutation">,
    listSharedWith: {} as FunctionReference<"query">,
  },
// ...
  categories: {
    listByProject: {} as FunctionReference<"query">,
    create: {} as FunctionReference<"mutation">,
    update: {} as FunctionReference<"mutation">,
    remove: {} as FunctionReference<"mutation">,
    reorder: {} as FunctionReference<"mutation">,
  },
  layers: {
    listByCategory: {} as FunctionReference<"query">,
    listByProject: {} as FunctionReference<"query">,
    create: {} as FunctionReference<"mutation">,
    updateWeight: {} as FunctionReference<"mutation">,
    updateName: {} as FunctionReference<"mutation">,
    markAsEffect: {} as FunctionReference<"mutation">,
    remove: {} as FunctionReference<"mutation">,
    getExclusions: {} as FunctionReference<"query">,
    createExclusion: {} as FunctionReference<"mutation">,
    deleteExclusion: {} as FunctionReference<"mutation">,
  },
// ...
} as const;